Implemented internal HP gain/loss cap at 32,000 (as per official), as the game's limited to a signed short (so it goes up to 32,767 before overflowing, which is bad).

+++ src/Ebenezer/MagicInstance.h
@@ -123,7 +123,7 @@ public:
 	void Type4Extend();
 
 	short GetMagicDamage(Unit *pTarget, int total_hit, int attribute);
-	short GetWeatherDamage(short damage, int attribute);
+	int32 GetWeatherDamage(int32 damage, int attribute);
 	void ReflectDamage(int32 damage, Unit * pTarget);
 	void ConsumeItem();
